requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.1
gitpython==3.1.40
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
httpx==0.27.0
